Docstrings and testing
